+++ BoraMorar/Cotacoes/DomainEvents/CotacaoStatusChangedDomainEvent.cs
@@ -1,4 +1,4 @@
 ﻿namespace BoraMorar.Cotacoes.DomainEvents
 {
-    public record CotacaoStatusChangedDomainEvent(int Id, CotacaoStatus Status, DateTime ChangedAt) : IDomainEvent { }
+    public record CotacaoStatusChangedDomainEvent(long Id, CotacaoStatus Status, DateTime ChangedAt) : IDomainEvent { }
 }
